+++ dbl/__init__.py
@@ -4,15 +4,15 @@
 DBL Python API Wrapper
 ~~~~~~~~~~~~~~~~~~~~~~
 A basic wrapper for the discordbots.org API.
-:copyright: (c) 2018 Francis Taylor & discordbots.org
+:copyright: (c) 2019 Francis Taylor & discordbots.org
 :license: MIT, see LICENSE for more details.
 """
 
 __title__ = 'dblpy'
 __author__ = 'Francis Taylor'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2018 Francis Taylor'
-__version__ = '0.1.6'
+__copyright__ = 'Copyright 2019 Francis Taylor'
+__version__ = '0.2'
 
 from collections import namedtuple
 
@@ -22,4 +22,4 @@ from .http import HTTPClient
 
 VersionInfo = namedtuple('VersionInfo', 'major minor micro releaselevel serial')
 
-version_info = VersionInfo(major=0, minor=1, micro=6, releaselevel='final', serial=0)
+version_info = VersionInfo(major=0, minor=2, micro=0, releaselevel='final', serial=0)