+++ dbl/http.py
@@ -3,7 +3,7 @@
 """
 The MIT License (MIT)
 
-Copyright (c) 2018 Francis Taylor
+Copyright (c) 2019 Francis Taylor
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -30,23 +30,19 @@ import json
 import logging
 import sys
 from datetime import datetime  # , timedelta
-from urllib.parse import urlencode
 from ratelimiter import RateLimiter
 
 import aiohttp
 
 from . import __version__
 from .errors import *
-#from .webhookserverhandler import WebHook
 
 log = logging.getLogger(__name__)
 
-
-@asyncio.coroutine
-def json_or_text(response):
+async def json_or_text(response):
     """Turns response into a properly formatted json or text object"""
-    text = response.text(encoding='utf-8')
-    if response.headers['content-type'] == 'application/json':
+    text = await response.text()
+    if response.headers['Content-Type'] == 'application/json; charset=utf-8':
         return json.loads(text)
     return text
 
@@ -77,7 +73,7 @@ class HTTPClient:
         self.user_agent = user_agent.format(__version__, sys.version_info, aiohttp.__version__)
 
 # TODO: better implementation of ratelimits
-# NOTE: current implementation doesn't (a) maintain state over restart and (b) wait an hour when a 429 is hit
+# NOTE: current implementation doesn't maintain state over restart
 
     async def request(self, method, url, **kwargs):
         """Handles requests to the API"""
@@ -87,46 +83,31 @@ class HTTPClient:
         async with rate_limiter:  # this works but doesn't 'save' over restart. need a better implementation.
 
             if not self.token:
-                raise UnauthorizedDetected('Unauthorized (status code: 401): No TOKEN provided')
+                raise UnauthorizedDetected('UnauthorizedDetected (status code: 401): No TOKEN provided')
 
             headers = {
                 'User-Agent': self.user_agent,
-                'Authorization': self.token
+                'Content-Type': 'application/json'
             }
 
             if 'json' in kwargs:
-                headers['Content-Type'] = 'application/json'
                 kwargs['data'] = to_json(kwargs.pop('json'))
 
             kwargs['headers'] = headers
+            headers['Authorization'] = self.token
 
-            # if not self._global_over.is_set():
-            # wait until the global lock is complete
-            #    await self._global_over.wait()
 
-            # await lock
-            # with MaybeUnlock(lock) as maybe_lock:
             for tries in range(5):
                 async with self.session.request(method, url, **kwargs) as resp:
                     log.debug('%s %s with %s has returned %s', method,
                               url, kwargs.get('data'), resp.status)
-                    text = await resp.text()
-                try:
+
                     data = await json_or_text(resp)
 
-                    # remaining = self._rl['quota']
-                    # if remaining == '0' and resp.status != 429:
-                    #     # we've depleted our current request quota
-                    #     delta = 60
-                    #
-                    #     log.warning(
-                    #         'A rate limit bucket has been exhausted (retry in: %s).', delta)
-                    #     maybe_lock.defer()
-                    #     self.loop.call_later(delta, lock.release)
 
                     if 300 > resp.status >= 200:
-                        data = json.loads(text)
                         return data
+
 
                     if resp.status == 429:  # we are being ratelimited
                         fmt = 'We are being rate limited. Retrying in %.2f seconds (%.3f minutes).'
@@ -152,6 +133,7 @@ class HTTPClient:
 
                         continue
 
+
                     if resp.status == 400:
                         raise HTTPException(resp, data)
                     elif resp.status == 401:
@@ -162,9 +144,6 @@ class HTTPClient:
                         raise NotFound(resp, data)
                     else:
                         raise HTTPException(resp, data)
-                finally:
-                    # clean-up just in case
-                    await resp.release()
             # We've run out of retries, raise.
             raise HTTPException(resp, data)
 
@@ -174,7 +153,7 @@ class HTTPClient:
     async def recreate(self):
         self.session = aiohttp.ClientSession(loop=self.session.loop)
 
-    async def post_server_count(self, bot_id, guild_count, shard_count, shard_no):
+    async def post_guild_count(self, bot_id, guild_count, shard_count, shard_no):
         if shard_count:
             payload = {
                 'server_count': guild_count,
@@ -185,11 +164,14 @@ class HTTPClient:
             payload = {
                 'server_count': guild_count
             }
-
         await self.request('POST', '{}/bots/{}/stats'.format(self.BASE, bot_id), json=payload)
 
-    async def get_server_count(self, bot_id):
-        '''Gets the server count of the given Bot ID'''
+    async def get_weekend_status(self):
+        '''Gets the weekend status from DBL'''
+        return await self.request('GET', '{}/weekend'.format(self.BASE))
+
+    async def get_guild_count(self, bot_id):
+        '''Gets the guild count of the given Bot ID'''
         return await self.request('GET', '{}/bots/{}/stats'.format(self.BASE, bot_id))
 
     async def get_bot_info(self, bot_id):
@@ -201,40 +183,21 @@ class HTTPClient:
                 resp[k] = None
         return resp
 
-    async def get_upvote_info(self, bot_id, onlyids, days):
-        '''Gets the upvote information of the given Bot ID'''
-        params = {
-            'onlyids': onlyids,
-            'days': days
-        }
-        return await self.request('GET', '{}/bots/{}/votes?'.format(self.BASE, bot_id) + urlencode(params))
+    async def get_upvote_info(self, bot_id):
+        '''Gets the last 1000 votes of the given Bot ID'''
+        return await self.request('GET', '{}/bots/{}/votes'.format(self.BASE, bot_id))
 
     async def get_bots(self, limit, offset):
-        '''Gets an object of all the bots on DBL'''
+        '''Gets an object of bots on DBL'''
+        if limit > 500:
+            limit = 50
         return await self.request('GET', '{}/bots?limit={}&offset={}'.format(self.BASE, limit, offset))
-
-    # async def search_bots(self, limit, offset, query):
-    #     return await self.request('GET', f'{self.BASE}/bots?limit={limit}&offset={offset}&search={query}')
 
     async def get_user_info(self, user_id):
         '''Gets an object of the user on DBL'''
         return await self.request('GET', '{}/users/{}'.format(self.BASE, user_id))
 
-    # async def initialize_webhooks(self, auth: str = None):
-    #     '''Initializes the webhook server'''
-    #     # setup webhook server
-    #     os.environ['HOOK_AUTH'] = str(auth)
-    #     os.environ['WEBHOOKS'] = True
-    #     await self.start_websocket_server()
-
-    @property
-    def bucket(self):
-        # the bucket is just method + path w/ major parameters
-        return '{0.method}:{0.url}'.format(self)
-
-
-@asyncio.coroutine
-def limited(until):
+async def limited(until):
     """Handles the message shown when we are ratelimited"""
     duration = int(round(until - time.time()))
     mins = duration / 60